Server: add authentication middleware and some permissions on tasks

+++ Server/src/routes/tasks.route.ts
@@ -1,10 +1,11 @@
 import * as express from "express";
 
 import { Task } from "../models/tasks";
+import { authMiddleware } from "../middlewares/auth";
 
 export const taskRouter = express.Router();
 
-taskRouter.use(express.json());
+taskRouter.use(authMiddleware);
 
 taskRouter.get("/", async (_req, res) => {
   Task.find()
@@ -43,6 +44,7 @@ taskRouter.post("/", async (req, res) => {
     title: receivedTask.title,
     priority: receivedTask.priority,
     interval: receivedTask.interval,
+    userId: receivedTask.userId,
   });
 
   task
@@ -65,39 +67,52 @@ taskRouter.post("/", async (req, res) => {
 taskRouter.put("/:id", async (req, res) => {
   const id = req?.params?.id;
   const task = req.body;
-
-  Task.updateOne({ _id: id }, task)
-    .then(() => {
-      console.log("Task updated :");
-      console.log(task);
-      res.status(201).json({
-        message: "Task updated successfully!",
-      });
-    })
-    .catch((error) => {
-      console.error(error);
+  Task.findOne({ _id: req.params.id }).then((task) => {
+    if (req.body.userId == task.userId) {
+      Task.updateOne({ _id: id }, task)
+        .then(() => {
+          console.log("Task updated :");
+          console.log(task);
+          res.status(201).json({
+            message: "Task updated successfully!",
+          });
+        })
+        .catch((error) => {
+          console.error(error);
+          res.status(400).json({
+            error: error,
+          });
+        });
+    } else {
       res.status(400).json({
-        error: error,
+        error: "You can't modify a task belonging to someone else",
       });
-    });
+    }
+  });
 });
 
 taskRouter.delete("/:id", async (req, res) => {
   Task.findOne({ _id: req.params.id }).then((task) => {
-    task
-      .deleteOne({ _id: req.params.id })
-      .then(() => {
-        console.log("Task deleted :");
-        console.log(task);
-        res.status(200).json({
-          message: "Task deleted successfully!",
+    if (req.body.userId == task.userId) {
+      task
+        .deleteOne({ _id: req.params.id })
+        .then(() => {
+          console.log("Task deleted :");
+          console.log(task);
+          res.status(200).json({
+            message: "Task deleted successfully!",
+          });
+        })
+        .catch((error) => {
+          console.error(error);
+          res.status(400).json({
+            error: error,
+          });
         });
-      })
-      .catch((error) => {
-        console.error(error);
-        res.status(400).json({
-          error: error,
-        });
+    } else {
+      res.status(400).json({
+        error: "You can't delete a task that belong to someone else",
       });
+    }
   });
 });
